+++ DesktopApplicationTemplate.UI/Navigation/INavigationHandler.cs
@@ -1,0 +1,12 @@
+using System.Threading.Tasks;
+using System.Windows.Controls;
+
+namespace DesktopApplicationTemplate.UI.Navigation
+{
+    public interface INavigationHandler
+    {
+        string ServiceType { get; }
+        Page CreateView(string defaultName);
+        Task AddServiceAsync(string name, object options);
+    }
+}